Implement new metaAllocation scheme using heap-based memory rather than end of file, avoiding need to copy allocators on file extend and preserving commit state across extension.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/JOURNAL_HA_BRANCH@3154 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/rwstore/FixedAllocator.java
@@ -44,7 +44,7 @@ public class FixedAllocator implements Allocator {
 	volatile private int m_freeBits;
 	volatile private int m_freeTransients;
 
-	private long m_diskAddr;
+	private int m_diskAddr;
 	int m_index;
 
 	protected boolean m_preserveSession = false;
@@ -80,11 +80,11 @@ public class FixedAllocator implements Allocator {
 		}
 	}
 
-	public long getDiskAddr() {
+	public int getDiskAddr() {
 		return m_diskAddr;
 	}
 
-	public void setDiskAddr(long addr) {
+	public void setDiskAddr(int addr) {
 		m_diskAddr = addr;
 	}
